FIX: Refactored upgrades to a more generic solution
