+++ intelmq/lib/upgrade/__init__.py
@@ -1,0 +1,10 @@
+# SPDX-FileCopyrightText: 2022 CERT.at GmbH <intelmq-team@cert.at>
+# SPDX-License-Identifier: AGPL-3.0-or-later
+#
+# -*- coding: utf-8 -*-
+from os.path import dirname, basename, isfile, join
+import glob
+
+
+modules = glob.glob(join(dirname(__file__), "*.py"))
+__all__ = sorted([basename(f)[:-3] for f in modules if isfile(f) and not basename(f).startswith('_')])
